Safari-611.1.10/LayoutTests/fast/selectors/attribute-list-with-whitespace-in-selector-expected.txt

When performing attribute matching for space separated values (e.g. [attr~=value]), the selector cannot match anything if the selector's "value" contains a white space. This test check this is true for all the definitions of whitespace of the spec.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll('[attr~=""]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~=" "]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="value "]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~=" value"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~=" value "]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="	"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="value	"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="	value"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="	value	"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="\\0a"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="\\0a "]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="value\\0a"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="\\0a value"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="\\0a value\\0a"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="\\0d"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="\\0d "]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="value\\0d"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="\\0d value"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="\\0d value\\0d"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="\\0c"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="\\0c "]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="value\\0c"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="\\0c value"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="\\0c value\\0c"]').length is 0
PASS hasElementStyledBySelector() is false
PASS document.querySelectorAll('[attr~="\\002003"]').length is 1
PASS hasElementStyledBySelector() is true
PASS document.querySelectorAll('[attr~="value\\002003"]').length is 1
PASS hasElementStyledBySelector() is true
PASS document.querySelectorAll('[attr~="\\002003 value"]').length is 1
PASS hasElementStyledBySelector() is true
PASS document.querySelectorAll('[attr~="\\002003 value\\002003"]').length is 1
PASS hasElementStyledBySelector() is true
PASS document.querySelectorAll('[attr~="\\003000"]').length is 1
PASS hasElementStyledBySelector() is true
PASS document.querySelectorAll('[attr~="value\\003000"]').length is 1
PASS hasElementStyledBySelector() is true
PASS document.querySelectorAll('[attr~="\\003000 value"]').length is 1
PASS hasElementStyledBySelector() is true
PASS document.querySelectorAll('[attr~="\\003000 value\\003000"]').length is 1
PASS hasElementStyledBySelector() is true
PASS successfullyParsed is true

TEST COMPLETE
